Reorganize root impls to a new module
